requirements.txt
openai-agents>=0.2.0
pandas>=2.2.0
python-dotenv>=1.0.0

